OpcodeDispatcher: Increase BitSize helper return vals to uint32_t
Pointed out by neobrain in #1362 that there are some cases where this
will overflow and not fit into 8 bits.

+++ External/FEXCore/Source/Interface/Core/OpcodeDispatcher.h
@@ -530,8 +530,8 @@ private:
 
   uint8_t GetDstSize(X86Tables::DecodedOp Op) const;
   uint8_t GetSrcSize(X86Tables::DecodedOp Op) const;
-  uint8_t GetDstBitSize(X86Tables::DecodedOp Op) const;
-  uint8_t GetSrcBitSize(X86Tables::DecodedOp Op) const;
+  uint32_t GetDstBitSize(X86Tables::DecodedOp Op) const;
+  uint32_t GetSrcBitSize(X86Tables::DecodedOp Op) const;
 
   template<unsigned BitOffset>
   void SetRFLAG(OrderedNode *Value) {
